+++ DTCD-StyleSystem/rollup.config.js
@@ -1,3 +1,5 @@
+import html from 'rollup-plugin-html';
+
 const watch = Boolean(process.env.ROLLUP_WATCH);
 
 const pluginName = 'StyleSystem';
@@ -5,7 +7,11 @@ const pluginName = 'StyleSystem';
 const outputFile = `${pluginName}.js`;
 const outputDirectory = watch ? `./../../DTCD/server/plugins/DTCD-${pluginName}` : `./build`;
 
-const plugins = [];
+const plugins = [
+  html({
+    include: '**/*.html',
+  }),
+];
 
 export default {
   input: `./src/${pluginName}.js`,