+++ DTCD-StyleSystem/src/base-components/components.js
@@ -1,0 +1,5 @@
+import BaseButton from './BaseButton/BaseButton';
+
+export default [
+  { name: 'base-button', baseClass: BaseButton }
+];